feat: refactor datetime options structure
